agents/pyproject.toml
[project]
name = "healthcare-assistant"
version = "0.1.0"
description = "Multi-agent virtual assistant system for healthcare using Strands Agents"
authors = [
    {name = "Healthcare Team", email = "healthcare@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Strands Agents Framework
    "strands-agents>=1.0.0",
    "strands-agents-tools>=0.2.13",
    # FastAPI and ASGI server
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    # AWS SDK and Services
    "boto3>=1.40.52",
    "botocore>=1.35.0",
    # Data Models and Validation
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    # HTTP and API
    "httpx>=0.25.0",
    # Async Support
    "aiofiles>=23.0.0",
    # Utilities
    "python-dateutil>=2.8.0",
    "typing-extensions>=4.8.0",
    "boto3>=1.40.62",
    "httpx-auth-awssigv4>=0.1.4"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "ruff>=0.0.290",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.ruff]
line-length = 88
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
